Remove `hostHeader = hostname` property from Host interceptor example
We are overriding the host header name from `host` to `hostname` in the example usage section. Due to this example users are overriding the header name too but still use the `%{host}` substitution as shown in the HDFS Sink section. This won't work for them.

This change removes this config line.